+++ changelog.md
@@ -1,10 +1,22 @@
 # Change Log
 
+## 0.9
+
+### Fixes
+
+- Increase timeouts for http requests
+- select role by alias
+- removed need for logon verb
+
+### Breaking change
+
+- removed version verb, use `--version` instead
+
 ## 0.8
 
 Complete re-write to make maintainance/enhancement easier.
 
-## New
+### New
 - config file now used in `~/.awsSts/config`
 - account alias is displayed and cached in `~/.awsSts/cache`
 
@@ -13,7 +25,7 @@ Complete re-write to make maintainance/enhancement easier.
 - Config can be updated even if aws cli configure has never been run
 - Ant required directories are created if required
 
-## Breaking changes
+### Breaking changes
 - Now defaults to 'profile' 'default'
 - flag `host` from 0.7 release is now `url`
 - must specify `logon` as the command to run